Refactor file location imports and component names
