+++ dashboard/app/(authenticated)/(main)/create/copilot/CreateCopilotProvider.tsx
