+++ dashboard/app/(authenticated)/(main)/create/copilot/DefineActionsStep.tsx
