+++ dashboard/app/(authenticated)/(main)/create/copilot/SetCopilotNameStep.tsx
